+optimize code

+++ server/index.js
@@ -2,6 +2,7 @@ const express = require("express");
 const cors = require("cors");
 const bodyParser = require("body-parser");
 const { PrismaClient } = require("@prisma/client");
+
 const prisma = new PrismaClient();
 const app = express();
 
@@ -11,36 +12,45 @@ app.use(bodyParser.json());
 
 app.get("/", async (req, res) => {
   try {
-    const displayAll = await prisma.taskTable.findMany({
-      include: {
-        tasks: true,
-      },
-    });
-    res.status(200).json(displayAll);
-    console.log(displayAll);
+    res.status(200).json(
+      await prisma.taskTable.findMany({
+        include: {
+          tasks: true,
+        },
+      })
+    );
   } catch (err) {
-    console.log(err);
+    res.status(500).json(err.message);
   }
 });
 
 app.post("/create/table", async (req, res) => {
   try {
     const { title } = req.body;
+
+    if (!title) return res.status(400).json({ message: "Data not provided!" });
+
     const newTable = await prisma.taskTable.create({
       data: {
         title: title,
       },
     });
-    res.status(200).redirect("/");
-    console.log(newTable);
+
+    if (newTable) {
+      res.status(200).redirect("/");
+    }
   } catch (err) {
-    console.log(err);
+    res.status(500).json(err.message);
   }
 });
 
 app.post("/create/task", async (req, res) => {
   try {
     const { title, description, author, tableId } = req.body;
+
+    if (!title || !description || !author || !tableId)
+      return res.status(400).json({ message: "Data not provided!" });
+
     const newTask = await prisma.tasks.create({
       data: {
         title: title,
@@ -49,16 +59,21 @@ app.post("/create/task", async (req, res) => {
         taskTableId: parseInt(tableId),
       },
     });
-    console.log(req.body.taskTableId);
-    res.status(200).json(newTask);
+
+    if (newTask) {
+      res.status(200).json(newTask);
+    }
   } catch (err) {
-    console.log(err);
+    res.status(500).json(err.message);
   }
 });
 
 app.delete("/delete/table", async (req, res) => {
   try {
     const { id } = req.body;
+
+    if (!id) return res.status(400).json({ message: "Data not provided!" });
+
     const deletedTable = await prisma.taskTable.delete({
       where: {
         id: id,
@@ -67,23 +82,32 @@ app.delete("/delete/table", async (req, res) => {
         tasks: true,
       },
     });
-    res.status(200).json(deletedTable);
+
+    if (deletedTable) {
+      res.status(200).json(deletedTable);
+    }
   } catch (err) {
-    console.log(err);
+    res.status(500).json(err.message);
   }
 });
 
 app.delete("/delete/task", async (req, res) => {
   try {
     const { id } = req.body;
+
+    if (!id) return res.status(400).json({ message: "Data not provided!" });
+
     const deletedTask = await prisma.tasks.delete({
       where: {
         id: id,
       },
     });
-    res.status(200).json(deletedTask);
+
+    if (deletedTask) {
+      res.status(200).json(deletedTask);
+    }
   } catch (err) {
-    console.log(err);
+    res.status(500).json(err.message);
   }
 });
